+++ ChangeLog.md
@@ -1,3 +1,7 @@
 # Changelog for gelbooru-scrape
 
-## Unreleased changes
+## 1.1.0.0
+ - Downloading all URLs from a newline-separated text file is now possible with the `-f` option
+ - Parallelism of downloading images was improved by not caring about the order they come in
+ - Parallelized download link extraction
+ - Made it possible to invoke the usage message through the `-h` option